Initial Version of the Library.
The library does what it needs to do.
Returns a frame from the calling stack; that should not be ignored. This allows
one to build a set of helper functions and classes, and still get the correct
context without the helpers getting in the way. One should be careful of how
many functions/packages are ignored, as it could lead to bloat.